// client/src/containers/RegisterModal/enums.ts
/*
 * enums
 * Enums allow us to organize a collection of related values. Think of them as
 * a class for values, wherein the value can only be a string , or number.
 *
 */

export enum fieldNames {
  userName = 'userName',
  email = 'email',
  password = 'password',
  confirmPassword = 'confirmPassword',
}